+++ src/main/java/net/therap/onlinestore/contoller/ItemController.java
@@ -2,14 +2,9 @@ package net.therap.onlinestore.contoller;
 
 import net.therap.onlinestore.entity.*;
 import net.therap.onlinestore.exception.IllegalAccessException;
-import net.therap.onlinestore.formatter.CategoryFormatter;
-import net.therap.onlinestore.formatter.TagFormatter;
-import net.therap.onlinestore.helper.AccessCheckHelper;
 import net.therap.onlinestore.helper.ItemHelper;
-import net.therap.onlinestore.service.CategoryService;
 import net.therap.onlinestore.service.FIleService;
 import net.therap.onlinestore.service.ItemService;
-import net.therap.onlinestore.service.TagService;
 import net.therap.onlinestore.validator.ItemValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.StringTrimmerEditor;
@@ -59,19 +54,10 @@ public class ItemController {
     private ItemService itemService;
 
     @Autowired
-    private CategoryService categoryService;
-
-    @Autowired
-    private TagService tagService;
+    private ItemHelper itemHelper;
 
     @Autowired
     private FIleService fIleService;
-
-    @Autowired
-    private CategoryFormatter categoryFormatter;
-
-    @Autowired
-    private TagFormatter tagFormatter;
 
     @Autowired
     private ItemValidator itemValidator;
@@ -82,13 +68,15 @@ public class ItemController {
     @InitBinder(ITEM)
     public void initBinder(WebDataBinder webDataBinder) {
         webDataBinder.registerCustomEditor(String.class, new StringTrimmerEditor(true));
-        webDataBinder.addCustomFormatter(categoryFormatter);
-        webDataBinder.addCustomFormatter(tagFormatter);
         webDataBinder.addValidators(itemValidator);
     }
 
     @GetMapping(ITEM_URL)
-    public String showItems(ModelMap modelMap) {
+    public String showItems(@SessionAttribute(value = ACTIVE_USER, required = false) User user,
+                            ModelMap modelMap) throws IllegalAccessException {
+
+        itemHelper.checkAccess(user, AccessType.VIEW_ALL);
+
         modelMap.put(ITEM_LIST, itemService.findAll());
         modelMap.put(NAV_ITEM, ITEM);
 
@@ -97,7 +85,9 @@ public class ItemController {
 
     @GetMapping(ITEM_DETAILS_URL)
     public String getItemDetails(@SessionAttribute(value = ORDER, required = false) Order order,
-                                 @RequestParam(ITEM_ID) int itemId, ModelMap modelMap) {
+                                 @RequestParam(ITEM_ID) int itemId,
+                                 ModelMap modelMap) {
+
         modelMap.put(ITEM, itemService.findById(itemId));
         modelMap.put(ORDER_ITEM, new OrderItem());
 
@@ -109,20 +99,26 @@ public class ItemController {
     }
 
     @GetMapping(ITEM_FORM_URL)
-    public String showItemForm(@SessionAttribute(value = ACTIVE_USER) User user,
+    public String showItemForm(@SessionAttribute(value = ACTIVE_USER, required = false) User user,
                                @RequestParam(value = ITEM_ID_PARAM, required = false) String itemId,
                                ModelMap modelMap) throws IllegalAccessException {
+
+        itemHelper.checkAccess(user, AccessType.FORM_LOAD);
+
         Item item = nonNull(itemId) ? itemService.findById(Integer.parseInt(itemId)) : new Item();
-        AccessCheckHelper.check(user, AccessType.LOAD_FORM, item);
         modelMap.put(ITEM, item);
-        setupReferenceDataItemForm(modelMap);
+        itemHelper.populateItemFromRefData(modelMap);
 
         return ITEM_FORM_VIEW;
     }
 
     @GetMapping(ITEM_CATEGORY_ID)
     @ResponseBody
-    public List<Item> getItemByCategoryId(@PathVariable(CATEGORY_ID) int categoryId) {
+    public List<Item> getItemByCategoryId(@SessionAttribute(value = ACTIVE_USER, required = false) User user,
+                                          @PathVariable(CATEGORY_ID) int categoryId) throws IllegalAccessException {
+
+        itemHelper.checkAccess(user, AccessType.READ);
+
         return itemService.findByCategory(categoryId);
     }
 
@@ -133,22 +129,23 @@ public class ItemController {
     }
 
     @PostMapping(ITEM_FORM_SAVE_URL)
-    public String saveOrUpdateItem(@SessionAttribute(ACTIVE_USER) User user,
+    public String saveOrUpdateItem(@SessionAttribute(value = ACTIVE_USER, required = false) User user,
                                    @Valid @ModelAttribute(ITEM) Item item,
                                    BindingResult bindingResult,
                                    ModelMap modelMap,
                                    SessionStatus sessionStatus,
                                    RedirectAttributes redirectAttributes) throws Exception {
 
+        itemHelper.checkAccess(user, AccessType.SAVE);
+
         if (bindingResult.hasErrors()) {
-            setupReferenceDataItemForm(modelMap);
+            itemHelper.populateItemFromRefData(modelMap);
 
             return ITEM_FORM_VIEW;
         }
 
         redirectAttributes.addFlashAttribute(SUCCESS, messageSource.getMessage(
                 (item.getId() == 0) ? "success.add" : "success.update", null, Locale.getDefault()));
-        AccessCheckHelper.check(user, AccessType.SAVE, item);
         itemService.saveOrUpdate(item);
         sessionStatus.setComplete();
 
@@ -156,20 +153,15 @@ public class ItemController {
     }
 
     @PostMapping(ITEM_DELETE_URL)
-    public String deleteItem(@SessionAttribute(ACTIVE_USER) User user,
+    public String deleteItem(@SessionAttribute(value = ACTIVE_USER, required = false) User user,
                              @RequestParam(ITEM_ID_PARAM) int itemId,
                              RedirectAttributes redirectAttributes) throws Exception {
-        AccessCheckHelper.check(user, AccessType.DELETE, itemService.findById(itemId));
+
+        itemHelper.checkAccess(user, AccessType.DELETE);
+
         itemService.delete(itemId);
         redirectAttributes.addFlashAttribute(SUCCESS, messageSource.getMessage("success.delete", null, Locale.getDefault()));
 
         return REDIRECT + ITEM_REDIRECT_URL;
-    }
-
-    private void setupReferenceDataItemForm(ModelMap modelMap) {
-        modelMap.put(CATEGORY_LIST, categoryService.findAll());
-        modelMap.put(ALL_TAG_LIST, tagService.findAll());
-        modelMap.put(AVAILABILITY_LIST, Availability.values());
-        modelMap.put(NAV_ITEM, ITEM);
     }
 }
